refactor(backend/services): Abstract custom errors into a separate file - this solves a circular dependency in tests
