+++ backend/src/services/GoogleAuth.ts
@@ -1,3 +1,4 @@
+import { Request } from "express";
 import { type Auth, google } from "googleapis";
 
 import { extractCode } from "../utils/extractQueryCode.js";
@@ -7,23 +8,9 @@ import {
   GOOGLE_CALLBACK_URL,
   GOOGLE_OAUTH_SCOPES,
 } from "../config/authConfig.js";
-
-export class GoogleAPIError extends Error {
-  constructor(message: string) {
-    // Call the constructor of the base class `Error`
-    // And set the error name to the custom error class name
-    super(message);
-    this.name = "GoogleAPIError";
-    // Set the prototype explicitly to maintain the correct prototype chain
-    Object.setPrototypeOf(this, GoogleAPIError.prototype);
-  }
-}
-
-export const throwGoogleError = (error: any) => {
-  throw new GoogleAPIError(
-    error instanceof Error ? error.message : String(error)
-  );
-};
+import { findOrCreateUserId } from "../controllers/index.js";
+import { throwGoogleError } from "./errors.js";
+import { GSession } from "../types/index.ts";
 
 class GoogleAuth {
   private authClient: Auth.OAuth2Client;
@@ -63,10 +50,10 @@ class GoogleAuth {
     });
   };
 
-  authenticate = async (req: any) => {
+  authenticate = async (req: Request) => {
     try {
       // Extract authorization code that will be exchanged for user tokens
-      const code = extractCode(req, req.session.googleAuthState);
+      const code = extractCode(req, (req.session as GSession).googleAuthState);
       // Because we are communicating directly with a Google server,
       // We can be confident that the token is valid
       const { tokens } = await this.getToken(code);
@@ -82,13 +69,12 @@ class GoogleAuth {
         // Get user email, name and picture
         const user = await this.getUserInfo();
         if (!user) return;
-        // interactWithDatabase(user);
-        const { email, name } = user;
-        return { email, name };
+        const userId = await findOrCreateUserId(user);
+        return { id: userId, displayName: user.displayName, email: user.email };
       }
     } catch (error) {
       console.error(error);
-      throwGoogleError(error);
+      throwGoogleError(String(error));
     }
   };
 
@@ -101,13 +87,13 @@ class GoogleAuth {
   getUserInfo = async () => {
     try {
       const { data } = await this.oauth2.userinfo.get();
-      const { email, name, picture } = data;
-      if (name && email) {
-        return { email, name, picture };
+      const { email, name: displayName, picture } = data;
+      if (displayName && email) {
+        return { email, displayName, picture };
       }
       return;
     } catch (error) {
-      throwGoogleError(error);
+      throwGoogleError(String(error));
     }
   };
 }
